requirements.txt
requests
beautifulsoup4
colorama
dnspython
python-whois
rich
tqdm
httpx
lxml
jinja2
asyncio
aiohttp
pyfiglet
